Moving failure analyzer out of the core module so that we can drop the Spring Boot dependency from graphql-dgs.
Moving failure analyzer out of the core module so that we can drop the Spring Boot dependency from graphql-dgs.
