+++ graphql-dgs-spring-boot-oss-autoconfigure/dependencies.lock
@@ -291,12 +291,6 @@
             ],
             "locked": "1.7.30"
         },
-        "org.springframework.boot:spring-boot": {
-            "firstLevelTransitive": [
-                "com.netflix.graphql.dgs:graphql-dgs"
-            ],
-            "locked": "2.3.6.RELEASE"
-        },
         "org.springframework.boot:spring-boot-dependencies": {
             "locked": "2.3.6.RELEASE"
         },
@@ -314,6 +308,12 @@
                 "com.netflix.graphql.dgs:graphql-dgs"
             ],
             "locked": "5.3.6.RELEASE"
+        },
+        "org.springframework:spring-context": {
+            "firstLevelTransitive": [
+                "com.netflix.graphql.dgs:graphql-dgs"
+            ],
+            "locked": "5.2.12.RELEASE"
         },
         "org.springframework:spring-framework-bom": {
             "locked": "5.2.12.RELEASE"
@@ -503,12 +503,6 @@
             ],
             "locked": "1.7.30"
         },
-        "org.springframework.boot:spring-boot": {
-            "firstLevelTransitive": [
-                "com.netflix.graphql.dgs:graphql-dgs"
-            ],
-            "locked": "2.3.6.RELEASE"
-        },
         "org.springframework.boot:spring-boot-dependencies": {
             "locked": "2.3.6.RELEASE"
         },
@@ -529,6 +523,12 @@
                 "com.netflix.graphql.dgs:graphql-dgs"
             ],
             "locked": "5.3.6.RELEASE"
+        },
+        "org.springframework:spring-context": {
+            "firstLevelTransitive": [
+                "com.netflix.graphql.dgs:graphql-dgs"
+            ],
+            "locked": "5.2.12.RELEASE"
         },
         "org.springframework:spring-framework-bom": {
             "locked": "5.2.12.RELEASE"
