--- graphql-dgs/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.diagnostics.FailureAnalyzer=com.netflix.graphql.dgs.internal.SchemaFailureAnalyzer
